Extract session from main glvis source file
