+++ glvis.cpp
@@ -11,7 +11,7 @@
 
 // GLVis - an OpenGL visualization server based on the MFEM library
 
-#include <limits>
+// #include <limits>
 #include <iostream>
 #include <fstream>
 #include <string>
@@ -35,6 +35,7 @@
 #endif
 
 #include <mfem.hpp>
+#include "lib/session.hpp"
 #include "lib/visual.hpp"
 #include "lib/window.hpp"
 #include "lib/script_controller.hpp"
@@ -46,15 +47,6 @@
 
 using namespace std;
 using namespace mfem;
-
-///////////////////////////////////////////////////////////////////////////////
-#ifdef NVTX_DBG_HPP
-#undef NVTX_COLOR
-#define NVTX_COLOR ::nvtx::kGold
-#include NVTX_DBG_HPP
-#else
-#define dbg(...)
-#endif
 
 const char *string_none    = "(none)";
 const char *string_default = "(default)";
@@ -78,105 +70,6 @@ thread_local GeometryRefiner GLVisGeometryRefiner;
 
 void PrintSampleUsage(ostream &out);
 
-class Session
-{
-   StreamCollection input_streams;
-   Window win;
-   std::thread handler;
-
-public:
-   Session(bool fix_elem_orient,
-           bool save_coloring,
-           string plot_caption,
-           bool headless)
-   {
-      win.data_state.fix_elem_orient = fix_elem_orient;
-      win.data_state.save_coloring = save_coloring;
-      win.plot_caption = plot_caption;
-      win.headless = headless;
-   }
-
-   Session(Window other_win)
-      : win(std::move(other_win))
-   { }
-
-   ~Session() = default;
-
-   Session(Session&& from) = default;
-   Session& operator= (Session&& from) = default;
-
-   inline DataState& GetState() { return win.data_state; }
-   inline const DataState& GetState() const { return win.data_state; }
-
-   void StartSession()
-   {
-      auto funcThread = [](Window w, StreamCollection is)
-      {
-         if (w.GLVisInitVis(std::move(is)))
-         {
-            w.GLVisStartVis();
-         }
-      };
-      handler = std::thread {funcThread,
-                             std::move(win), std::move(input_streams)};
-      handler.detach();
-   }
-
-   bool StartSavedSession(std::string stream_file)
-   {
-      unique_ptr<ifstream> ifs(new ifstream(stream_file));
-      if (!(*ifs))
-      {
-         cout << "Can not open stream file: " << stream_file << endl;
-         return false;
-      }
-      string data_type;
-      *ifs >> data_type >> ws;
-      StreamReader reader(win.data_state);
-      reader.ReadStream(*ifs, data_type);
-      input_streams.emplace_back(std::move(ifs));
-
-      StartSession();
-      return true;
-   }
-
-   int StartStreamSession(std::unique_ptr<mfem::socketstream> &&stream,
-                          const std::string &data_type)
-   {
-      StreamReader reader(win.data_state);
-      int ierr = reader.ReadStream(*stream, data_type);
-      if (ierr) { return ierr; }
-      input_streams.emplace_back(std::move(stream));
-
-      StartSession();
-      return 0;
-   }
-
-   int StartStreamSession(StreamCollection &&streams)
-   {
-      dbg();
-      StreamReader reader(win.data_state);
-      int ierr = reader.ReadStreams(streams);
-      if (ierr) { return ierr; }
-      input_streams = std::move(streams);
-
-      StartSession();
-      return 0;
-   }
-
-   int StartSerialStreamSession(std::unique_ptr<std::istream> &&stream,
-                                const std::string &data_type)
-   {
-      dbg("data_type: '{}'", data_type);
-      StreamReader reader(win.data_state);
-      int ierr = reader.ReadStream(*stream, data_type);
-      if (ierr) { dbg("❌ ERROR ❌"); return ierr; }
-
-      input_streams.emplace_back(std::move(stream));
-      StartSession();
-      return 0;
-   }
-};
 
 void GLVisServer(int portnum, bool save_stream, bool fix_elem_orient,
                  bool save_coloring, string plot_caption, bool headless = false)
@@ -367,44 +260,6 @@ void GLVisServer(int portnum, bool save_stream, bool fix_elem_orient,
       }
       current_sessions.emplace_back(std::move(new_session));
    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-void *GLVisLibGetWindow()
-{
-   static Window win;
-   return (void*) &win;
-}
-
-///////////////////////////////////////////////////////////////////////////////
-int GLVisLibWindow(void *win_ptr, bool fix_elem_orient,
-                   bool save_coloring, bool headless,
-                   const std::string &plot_caption,
-                   std::unique_ptr<std::istream> &&stream,
-                   const std::string &data_type)
-{
-   dbg();
-   // std::this_thread::sleep_for(std::chrono::milliseconds(500));
-
-   dbg("Main Window structure");
-   auto *win = static_cast<Window*>(win_ptr);
-
-   std::vector<Session> current_sessions;
-
-   dbg("Get main thread");
-   GetMainThread(win->headless);
-
-   dbg("StartStreamSession");
-   Session new_session(fix_elem_orient, save_coloring, plot_caption, headless);
-
-   new_session.StartSerialStreamSession(std::move(stream), data_type);
-   current_sessions.emplace_back(std::move(new_session));
-
-   dbg("Starting message loop in main thread");
-   MainThreadLoop();
-
-   dbg("EXIT_SUCCESS");
-   return EXIT_SUCCESS;
 }
 
 int main (int argc, char *argv[])
